+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - ContentCheck - allow targeting of root object or array (#334)
 - Tests - support bracket notation for targeted keys (#333)
 - Fuzzing - support JSON variants as content-type (#332)
+- Fuzzing - added support for allOf, anyOf, oneOf schemas (#335)
 
 ## v1.16.1 - (2022-04-13)
 - Set the 'logAssignVariables' option via 'cliOptions' to toggle console output for assigned variables